+++ packages/plugins/type-designer/src/headless/stores/type-elements/ref-crud-operation.helper.ts
@@ -19,9 +19,11 @@ function getRefAttributes(params: {
 }) {
 	return {
 		[REF_FAMILY.generalEquipment]: () => ({
+			virtual: 'false',
 			templateUuid: params.typeId
 		}),
 		[REF_FAMILY.conductingEquipment]: () => ({
+			virtual: 'false',
 			templateUuid: params.typeId
 		}),
 		[REF_FAMILY.function]: () => ({