+++ packages/plugins/type-designer/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 27-02-2025
+
+### Added
+
+- Virtual attribute is added on equipment creation (default to 'false')
+- Sidebar close on type deletion
+
 ## [3.1.0] - 26-02-2025
 
 ### Added